requirements.txt
keybert
sudachipy
scikit-learn
sentence-transformers
transformers
pyyaml
anthropic
python-dotenv 